+++ source/bin/rocprof-sys-sample/impl.cpp
@@ -339,20 +339,23 @@ parse_args(int argc, char** argv, std::vector<char*>& _env)
                                                original_envs);
         });
     parser
-        .add_argument({ "-T", "--trace" }, "Generate a detailed trace (perfetto output)")
+        .add_argument(
+            { "-T", "--trace" },
+            "Generate a detailed trace with deferred trace generation (perfetto output)")
         .max_count(1)
         .action([&](parser_t& p) {
-            rocprofsys::common::update_env(_env, "ROCPROFSYS_TRACE", p.get<bool>("trace"),
-                                           update_mode::REPLACE, ":", updated_envs,
-                                           original_envs);
+            rocprofsys::common::update_env(_env, "ROCPROFSYS_TRACE_CACHED",
+                                           p.get<bool>("trace"), update_mode::REPLACE,
+                                           ":", updated_envs, original_envs);
         });
     parser
-        .add_argument({ "--trace-cached" },
-                      "Generate a detailed trace (perfetto output) from cached data ")
+        .add_argument(
+            { "-L", "--trace-legacy" },
+            "Generate a detailed trace with direct mode (perfetto output, legacy)")
         .max_count(1)
         .action([&](parser_t& p) {
             rocprofsys::common::update_env(
-                _env, "ROCPROFSYS_TRACE_CACHED", p.get<bool>("trace-cached"),
+                _env, "ROCPROFSYS_TRACE_LEGACY", p.get<bool>("trace-legacy"),
                 update_mode::REPLACE, ":", updated_envs, original_envs);
         });
     parser