+++ tests/rocprof-sys-thread-limit-tests.cmake
@@ -32,7 +32,8 @@ endif()
 
 set(_thread_limit_environment
     "${_base_environment}"
-    "ROCPROFSYS_TRACE=ON"
+    "ROCPROFSYS_TRACE_LEGACY=OFF"
+    "ROCPROFSYS_TRACE_CACHED=ON"
     "ROCPROFSYS_PROFILE=ON"
     "ROCPROFSYS_COUT_OUTPUT=ON"
     "ROCPROFSYS_USE_SAMPLING=ON"